Added loader which loads core components before the main software
