+++ package.json
@@ -32,7 +32,7 @@
     "lint:check": "eslint . --ext .ts",
     "build:check": "tsc -p tsconfig.json",
     "build": "rimraf dist && node build.js",
-    "start": "node ./dist/index.js"
+    "start": "node ./dist/loader.js"
   },
   "devDependencies": {
     "@sentry/cli": "^2.31.2",
